clean up and helpr utils
